% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetImInfo.R
\name{getImInfo}
\alias{getImInfo}
\title{Get Information about images}
\usage{
getImInfo(PlateID, SQL.use = F, SERVER = "MDCStore", DB, PlateLoc = "...",
  TimeCourse = F)
}
\description{
Get Information about images
}
\keyword{internal}
